Improve pgpVerifySelf() API
The pgpVerifySelf() API was rather clumsy.  The type of the data packet
should determine the type of its self-signature, not the other way
around.  This makes detecting wrongly-typed self-signatures much
simpler.

Now that pgpPrtParamsSubkeys() checks the type of self-signatures,
subkey revocation signatures can be checked too.  This allows importing
keys that have revoked subkeys.  Also add a bounds-check in
pgpVerifySelf() in case the index passed is wrong.