Restrict packets following a subkey
According to RFC 4880, subkeys are supposed to only be followed by
binding signatures, revocation signatures, and other subkeys.  In
practice, the most likely reason to get something else is a revocation
signature of a type that RPM does not implement.  This could include a
revocation of a subkey binding signature or a revocation made by a
designated revoker.  In this case, it is not safe to import the key, as
RPM might accept a signature made by a subkey whose secret portion has
been compromised.  Therefore, return an error in this case.